Stop exiting when the player stops.
This commit started off as an attempt to fix the handling of player
state changes; `mpdpopm' was failing whenever the player was
stopped. While doing so, I wound up re-writing large swaths
of code. The current status is akin to having your computer
disassembled on the living room floor; it boots up, but there are
screws everywhere & disconnected cables on top of it all.