[#6] Update the test scripts for Arch packing.
As part of issue #6, I'm changing the way I make an Arch package to
publish a few PKGBUILDs (-bin, -git, &c). This patchset updates a
few test scripts.

Also, add LICENSE to the distribution & fix the Debian package build.
Somehow, the name got out of sync; I was missing a '-1'.