Move from `log4rs` to `tracing` for logging.
This patch moves mpdpopm over to the `tracing` crate from `log4rs`.
It also clears out a few more lints exposed by the new signoff
script.

This commit should perhaps be squashed with the next; since that
commit will involve moving over to `snafu` for error handling,
I handled several points-of-failure in this commit with `unwrap()`.